requirements.txt
opencv-python==4.6.0.66
dill==0.3.8
onnx==1.16.1
onnxruntime==1.18.0
streamlit==1.30.0
pillow==10.2.0
pandas==2.1.4
numpy==1.26.4
scikit-image
shapely==2.0.5
paddleocr==2.8.1
paddlepaddle==2.6.1
protobuf==3.20.2
fastAPI
uvicorn
loguru